backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-environ==0.10.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.18
phonenumberslite==8.13.18
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
